added search and partial filter functionality
Searching is broken by filters at the moment.

Recipes can be serached for by their name, description or author.

Recipes can be filtered by total preparation and cooking time, based on the information supplied with the recipes. For recipes without this information I need to finish developing a regex to work it out.